requirements.txt
h5py==2.8.0
llvmlite==0.25.0+0.g9af98a6.dirty
matplotlib==2.1.0
numba==0.40.0
numpy==1.17.0
Pillow==5.2.0
scikit-learn==0.20.0
scipy==1.3.1
torch==0.4.1
torchfile==0.1.0
torchvision==0.2.1
tqdm==4.28.1